feat: Add unit tests for compressing ipv6 and move ip-addresses.ts from test to utils
