+++ frontend/src/ts/utils/misc.ts
@@ -673,72 +673,4 @@ export function updateTitle(title?: string): void {
   }
 }
 
-export function compressIpv6(ip: string): string {
-  let ipArray = ip.split(":");
-  let newIp = "";
-
-  // Replace the leading zeros using regex
-  ipArray = ipArray.map((word) => word.replace(/^0+/gm, ""));
-
-  // Find the longest series of zeros
-  let longestSeriesNum = 0,
-    longestStartIndex = -1,
-    longestEndIndex = -1;
-  let currentSeriesNum = 0,
-    currentStartIndex = -1,
-    currentEndIndex = -1;
-
-  for (let i = 0; i < ipArray.length; i++) {
-    if (ipArray[i] === "") {
-      if (currentStartIndex === -1) currentStartIndex = i;
-      currentSeriesNum++;
-    } else {
-      // The last empty index was the previous one
-      currentEndIndex = i - 1;
-
-      if (longestSeriesNum < currentSeriesNum) {
-        longestSeriesNum = currentSeriesNum;
-        longestStartIndex = currentStartIndex;
-        longestEndIndex = currentEndIndex;
-      }
-
-      // Reset the values
-      currentSeriesNum = 0;
-      currentStartIndex = -1;
-      currentEndIndex = -1;
-    }
-  }
-
-  if (longestSeriesNum < currentSeriesNum) {
-    longestSeriesNum = currentSeriesNum;
-    longestStartIndex = currentStartIndex;
-    longestEndIndex = ipArray.length - 1;
-  }
-
-  // Fill in zeroes in all places except within and including the start and end indexes
-  // In those indexes, use a ':'
-
-  ipArray = ipArray.map((word, index) => {
-    if (word === "") {
-      // If the index is within those indexes and that this is not a single 16bit zero field
-      if (
-        index >= longestStartIndex &&
-        index <= longestEndIndex &&
-        longestStartIndex !== longestEndIndex
-      )
-        return ":";
-      else return "0";
-    }
-
-    return word;
-  });
-
-  newIp = ipArray.join(":");
-  // Replace multiple colons with a double colon.
-  // This should occur only once
-  newIp = newIp.replace(/:{3,}/gm, "::");
-
-  return newIp;
-}
-
 // DO NOT ALTER GLOBAL OBJECTSONSTRUCTOR, IT WILL BREAK RESULT HASHES
